+++ src/main/java/com/fitlogtimer/dto/ExerciseSetOutDTO.java
@@ -2,7 +2,7 @@ package com.fitlogtimer.dto;
 
 public record ExerciseSetOutDTO(
         Long id,
-        Long exercise_id,
+        int exercise_id,
         double weight,
         int repNumber,
         boolean isMax,
